feat: more folder restructuring
