--- webex_assistant_skills_sdk/api/shared/models/directives/__init__.py
@@ -1,22 +1,0 @@
-from webex_assistant_skills_sdk.api.shared.models.directives.asr_hint import ASRHint
-from webex_assistant_skills_sdk.api.shared.models.directives.assistant_event import AssistantEvent
-from webex_assistant_skills_sdk.api.shared.models.directives.clear_web_view import ClearWebView
-from webex_assistant_skills_sdk.api.shared.models.directives.display_web_view import DisplayWebView
-from webex_assistant_skills_sdk.api.shared.models.directives.listen import Listen
-from webex_assistant_skills_sdk.api.shared.models.directives.reply import Reply
-from webex_assistant_skills_sdk.api.shared.models.directives.sleep import Sleep
-from webex_assistant_skills_sdk.api.shared.models.directives.speak import Speak
-from webex_assistant_skills_sdk.api.shared.models.directives.ui_hint import UIHint
-
-
-__all__ = [
-    'ASRHint',
-    'AssistantEvent',
-    'ClearWebView',
-    'DisplayWebView',
-    'Listen',
-    'Reply',
-    'Sleep',
-    'Speak',
-    'UIHint',
-]
